+++ NEWS.rst
@@ -5,3 +5,16 @@ Version 0.5: April 2023
 ***********************
 
 First public version.
+
+Version 0.5.1: April 2023
+*************************
+
+Minor fixes.
+
+First public version.
+
+Version 0.5.2: May 2023
+***********************
+
+Added dig_sequence command, which allows a sequence stored locally or in Genbank to be searched for a specific
+allele. The closest match will be annotated.